requirements.txt
aiofiles==22.1.0 ; python_full_version == "3.9.16"
aiogram==3.0.0b6 ; python_full_version == "3.9.16"
aiohttp==3.8.3 ; python_full_version == "3.9.16"
aiomysql==0.1.1 ; python_full_version == "3.9.16"
aiosignal==1.3.1 ; python_full_version == "3.9.16"
alembic==1.9.2 ; python_full_version == "3.9.16"
anyio==3.6.2 ; python_full_version == "3.9.16"
async-timeout==4.0.2 ; python_full_version == "3.9.16"
attrs==22.2.0 ; python_full_version == "3.9.16"
certifi==2022.12.7 ; python_full_version == "3.9.16"
cffi==1.15.1 ; python_full_version == "3.9.16"
charset-normalizer==2.1.1 ; python_full_version == "3.9.16"
click==8.1.3 ; python_full_version == "3.9.16"
colorama==0.4.6 ; sys_platform == "win32" and python_full_version == "3.9.16" or platform_system == "Windows" and python_full_version == "3.9.16"
cryptography==39.0.0 ; python_full_version == "3.9.16"
fastapi==0.89.1 ; python_full_version == "3.9.16"
frozenlist==1.3.3 ; python_full_version == "3.9.16"
greenlet==2.0.2 ; platform_machine == "aarch64" and python_full_version == "3.9.16" or platform_machine == "ppc64le" and python_full_version == "3.9.16" or platform_machine == "x86_64" and python_full_version == "3.9.16" or platform_machine == "amd64" and python_full_version == "3.9.16" or platform_machine == "AMD64" and python_full_version == "3.9.16" or platform_machine == "win32" and python_full_version == "3.9.16" or platform_machine == "WIN32" and python_full_version == "3.9.16"
h11==0.14.0 ; python_full_version == "3.9.16"
httptools==0.5.0 ; python_full_version == "3.9.16"
idna==3.4 ; python_full_version == "3.9.16"
jinja2==3.1.2 ; python_full_version == "3.9.16"
magic-filter==1.0.9 ; python_full_version == "3.9.16"
mako==1.2.4 ; python_full_version == "3.9.16"
markupsafe==2.1.2 ; python_full_version == "3.9.16"
multidict==6.0.4 ; python_full_version == "3.9.16"
pycparser==2.21 ; python_full_version == "3.9.16"
pydantic==1.10.4 ; python_full_version == "3.9.16"
pymysql==1.0.2 ; python_full_version == "3.9.16"
python-dotenv==0.21.1 ; python_full_version == "3.9.16"
pyyaml==6.0 ; python_full_version == "3.9.16"
redis==4.4.2 ; python_full_version == "3.9.16"
sniffio==1.3.0 ; python_full_version == "3.9.16"
sqlalchemy==2.0.1 ; python_full_version == "3.9.16"
starlette==0.22.0 ; python_full_version == "3.9.16"
typing-extensions==4.4.0 ; python_full_version == "3.9.16"
uvicorn[standard]==0.20.0 ; python_full_version == "3.9.16"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_full_version == "3.9.16"
watchfiles==0.18.1 ; python_full_version == "3.9.16"
websockets==10.4 ; python_full_version == "3.9.16"
yarl==1.8.2 ; python_full_version == "3.9.16"
